add create and clear dict when ends
